Move WorkThumbnail.svelte into a WorkThumbnail/ directory to match component convention
Other components (GlobalSideNav, Icon, CommentTree, Footer, etc.)
already live in their own directory alongside their story file.
WorkThumbnail.svelte was still a flat file at the lib root, so this
brings it in line, and updates all seven importers accordingly.
