+++ frontend/src/lib/WorkThumbnail/WorkThumbnail.stories.ts
