+++ frontend/src/lib/WorkThumbnail/WorkThumbnail.svelte
